+++ src/Commands/RapydMakeEditCommand.php
@@ -12,7 +12,7 @@ use Zofe\Rapyd\Utilities\StrReplacer;
 
 class RapydMakeEditCommand extends RapydMakeBaseCommand
 {
-    public $signature = 'rpd:make:edit {component} {model} {--module=} {--table=} {--fields=}';
+    public $signature = 'rpd:make:edit {component} {model} {--module=} {--table=} {--fields=} {--increments}';
 
     public $description = 'rapyd command to generate DataEdit component';
 
@@ -39,8 +39,9 @@ class RapydMakeEditCommand extends RapydMakeBaseCommand
         $componentName = $component;
         $component_name = Str::snake($componentName);
 
-        $this->createModuleConfig();
         $this->createModel($model);
+        $this->createModuleConfig();
+        $this->createAuthorizationAndLimit($model);
 
         $this->comment('generate '.$component.' for model '.$model);